+++ mywords-flutter/lib/common/global.dart
@@ -4,9 +4,9 @@ class Global {
   // 需要在main函数中初始化，且在handler.initLib()之后。 如果有更新。本地缓存也需要更新。
   static Map<String, int> allKnownWordsMap = {}; //TODO: 本地缓存 梳理下都有哪些地方会更新这个数据
   static String parseVersion = '';
-   static int webDictRunPort = 0;
-   static int webOnlinePort = 0;
-  static const version = "3.0.0";
+  static int webDictRunPort = 0;
+  static int webOnlinePort = 0;
+  static const version = "3.0.1";
   static String goBuildInfoString = "";
 
   static Map<int, int> levelDistribute(List<String> words) {
@@ -20,14 +20,14 @@ class Global {
 
   static Future<void> init() async {
     handler.getDefaultDictId();
-     parseVersion = await handler.parseVersion();
+    parseVersion = await handler.parseVersion();
     goBuildInfoString = await handler.goBuildInfoString();
     final respData = await handler.allKnownWordsMap();
     if (respData.code != 0) {
       throw Exception("allKnownWordsMap error: ${respData.message}");
     }
+    allKnownWordsMap = respData.data ?? {};
     webDictRunPort = await handler.webDictRunPort();
     webOnlinePort = await handler.webOnlinePort();
-    allKnownWordsMap = respData.data ?? {};
   }
 }