+++ mywords-flutter/lib/common/read_message.dart
@@ -1,4 +1,8 @@
+import 'dart:convert';
 import 'dart:isolate';
+
+import 'package:mywords/common/global.dart';
+import 'package:mywords/common/queue.dart';
 
 import '../libso/handler.dart';
 import '../util/util.dart';
@@ -42,12 +46,29 @@ void isolateLoopReadMessage() async {
       final separatorIdx = response.indexOf(":");
       final code = int.parse(response.substring(0, separatorIdx));
       final content = response.substring(separatorIdx + 1);
+      myPrint("code: $code, content: $content");
       switch (code) {
         case 0:
           myPrint("receive error message: $content");
-
           continue;
         case 1:
+          //更新单词列表
+          myPrint("更新单词列表");
+          // ["result",1]
+          List<dynamic> param = jsonDecode(content);
+          if (param.length >= 2) {
+            if (param[0] is String && param[1] is int) {
+              final word = param[0];
+              final level = param[1];
+              if (level == 0) {
+                Global.allKnownWordsMap.remove(word);
+                produceEvent(EventType.updateKnownWord);
+              } else {
+                Global.allKnownWordsMap[word] = level;
+                produceEvent(EventType.updateKnownWord);
+              }
+            }
+          }
           break;
         case 2:
           break;